impl tx_history for WASM and refactoring
